icon-loader: load SVG icons at max scale of any usable output
- Add a new function to get the maximum scale of all usable outputs
- Pass the maximum output scale through to img_svg_load(), which
  ultimately calls cairo_surface_set_device_scale() before rendering

+++ include/icon-loader.h
@@ -7,6 +7,6 @@ struct server;
 void icon_loader_init(struct server *server);
 void icon_loader_finish(struct server *server);
 struct lab_data_buffer *icon_loader_lookup(struct server *server,
-	const char *app_id, int size, int scale);
+	const char *app_id, int size, float scale);
 
 #endif /* LABWC_ICON_LOADER_H */
